[Sequence/StartServiceUDPDataReceiveSequence] Moved ConfigParser under JsonParser
As Config Parser contains JSON parsing, and all JSON releated activities should be performed under jsonparser module, it is moved from SensorMoniteringHubManager module to JsonParser module
